requirement.txt
Django==1.11.7
django-enumerify==0.5.1
django-filter==1.0.2
djangorestframework==3.7.3
pkg-resources==0.0.0
pytz==2017.3
